# docs/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiEnvelope_SearchData_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[SearchData]",
        "type": "object"
      },
      "ApiEnvelope_SearchInspectData_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchInspectData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[SearchInspectData]",
        "type": "object"
      },
...
      "ApiEnvelope_TargetPostsData_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TargetPostsData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[TargetPostsData]",
        "type": "object"
      },
      "ApiEnvelope_TargetSocialTimelineData_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TargetSocialTimelineData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[TargetSocialTimelineData]",
        "type": "object"
      },
      "ApiEnvelope_TokenCaseData_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenCaseData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[TokenCaseData]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SearchData": {
        "additionalProperties": true,
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "additionalProperties": true,
            "title": "Page",
            "type": "object"
          },
          "query": {
            "additionalProperties": true,
            "title": "Query",
            "type": "object"
          },
          "target_candidates": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Target Candidates",
            "type": "array"
          }
        },
        "required": [
          "query",
          "page"
        ],
        "title": "SearchData",
        "type": "object"
      },
      "SearchInspectData": {
        "additionalProperties": true,
        "properties": {
          "ambiguous_result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ambiguous Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "query": {
            "additionalProperties": true,
            "title": "Query",
            "type": "object"
          },
          "resolver": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolver"
          },
          "token_result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Result"
          },
          "topic_result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Result"
          }
        },
        "required": [
          "query"
        ],
        "title": "SearchInspectData",
        "type": "object"
      },
...
      "TargetPostsData": {
        "additionalProperties": true,
        "properties": {
          "has_more": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has More"
          },
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "query": {
            "additionalProperties": true,
            "title": "Query",
            "type": "object"
          },
          "returned_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returned Count"
          },
          "score_window": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Window"
          },
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count"
          }
        },
        "required": [
          "query"
        ],
        "title": "TargetPostsData",
        "type": "object"
      },
      "TargetSocialTimelineData": {
        "additionalProperties": true,
        "properties": {
          "authors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Authors",
            "type": "array"
          },
          "buckets": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Buckets",
            "type": "array"
          },
          "cascade": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cascade"
          },
          "has_more": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has More"
          },
          "market_candles": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Candles"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "posts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Posts",
            "type": "array"
          },
          "query": {
            "additionalProperties": true,
            "title": "Query",
            "type": "object"
          },
          "returned_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returned Count"
          },
          "stages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Stages",
            "type": "array"
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "query"
        ],
        "title": "TargetSocialTimelineData",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Parallax",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/search": {
      "get": {
        "operationId": "search_api_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": "",
              "title": "Cursor",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "24h",
              "title": "Window",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_SearchData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "api"
        ]
      }
    },
    "/api/search/inspect": {
      "get": {
        "operationId": "search_inspect_api_search_inspect_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "24h",
              "title": "Window",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_SearchInspectData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Inspect",
        "tags": [
          "api"
        ]
      }
    },
...
    "/api/target-posts": {
      "get": {
        "operationId": "target_posts_api_target_posts_get",
        "parameters": [
          {
            "in": "query",
            "name": "target_type",
            "required": false,
            "schema": {
              "default": "",
              "title": "Target Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "Target Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "5m",
              "title": "Window",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "range",
            "required": false,
            "schema": {
              "default": "current_window",
              "title": "Range",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "recent",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": "",
              "title": "Cursor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_TargetPostsData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Target Posts",
        "tags": [
          "api"
        ]
      }
    },
    "/api/target-social-timeline": {
      "get": {
        "operationId": "target_social_timeline_api_target_social_timeline_get",
        "parameters": [
          {
            "in": "query",
            "name": "target_type",
            "required": false,
            "schema": {
              "default": "",
              "title": "Target Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "Target Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "1h",
              "title": "Window",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": "",
              "title": "Cursor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_TargetSocialTimelineData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Target Social Timeline",
        "tags": [
          "api"
        ]
      }
    },
    "/api/token-case": {
      "get": {
        "operationId": "token_case_api_token_case_get",
        "parameters": [
          {
            "in": "query",
            "name": "target_type",
            "required": false,
            "schema": {
              "default": "",
              "title": "Target Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "Target Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": "1h",
              "title": "Window",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "all",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "posts_limit",
            "required": false,
            "schema": {
              "default": 24,
              "title": "Posts Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_TokenCaseData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Token Case",
        "tags": [
          "api"
        ]
      }
    },
...
  }
}